+++ Changelog.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
  - ngFormioGrid display being broken
  - View not being updated when inner submission fields are changed
+ - Buttons with 'event' action firing events twice
+ - Watchers not firing when input is changed on the form
 
 ## 2.19.5
 ### Changed